extract getFiles to reusable file
